Fix inconsistencies in renaming files
